fixing users query

+++ src/Resolver/Skill.resolver.js
@@ -1,21 +1,49 @@
 import { Skill } from "../Models/Users/Skills.model.js";
+import { User } from "../Models/Users/Users.model.js";
 
 const skillResolvers = {
   Query: {
-    skills: async () => await Skill.find(),
-    skill: async (parent, args) => await Skill.findById(args.id),
+    skills: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in');
+      }
+      return await Skill.find()
+    },
+    skill: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in');
+      }
+      return await Skill.findById(args.id)
+    }
   },
   Mutation: {
-    createSkill: async (parent, args) => {
+    createSkill: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in');
+      }
       const newSkill = new Skill(args);
+      await User.findByIdAndUpdate(context.user.id, { $push: { skills: newSkill._id } });
       return await newSkill.save();
     },
-    updateSkill: async (parent, args) => {
+    updateSkill: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in');
+      }
       const updatedSkill = await Skill.findByIdAndUpdate(args.id, args, { new: true });
+      if (!updatedSkill) {
+        throw new Error('Skill not found');
+      }
       return updatedSkill;
     },
-    deleteSkill: async (parent, args) => {
+    deleteSkill: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('You must be logged in');
+      }
       const deletedSkill = await Skill.findByIdAndRemove(args.id);
+      if (!deletedSkill) {
+        throw new Error('Skill not found');
+      }
+      await User.findByIdAndUpdate(context.user.id, { $pull: { skills: args.id } });
       return deletedSkill;
     },
   },
